Query eQTL Catalogue:tabix/echotabix

Query eQTL Catalogue datasets by region with tabix or echotabix. Faster alternative to REST API.

fetch_tabix(
  dataset_id,
  query_granges,
  query_method = c("rsamtools", "variantannotation", "conda", "seqminer"),
  quant_method = "ge",
  nThread = 1,
  conda_env = "echoR_mini",
  verbose = TRUE
)

Source

eQTL Catalogue blocking tabix requests

data("meta"); query_granges <- echotabix::construct_query(query_dat = echodata::BST1) qtl.subset <- catalogueR:::fetch_tabix(unique_label=meta$unique_label[2], query_granges=query_granges)

Arguments

query_granges

GRanges object to be used for querying the target_path file. Alternatively, can be variant-level summary statistics to be converted into a GRanges object by construct_query.

query_method

Method used for querying. See query for available options.

quant_method

eQTL Catalogue actually contains more than just eQTL data. For each dataset, the following kinds of QTLs can be queried:

gene expression QTL

quant_method="ge" (default) or quant_method="microarray", depending on the dataset. catalogueR will automatically select whichever option is available.

exon expression QTL

*under construction* quant_method="ex"

transcript usage QTL

*under construction* quant_method="tx"

promoter, splice junction and 3' end usage QTL

*under construction* quant_method="txrev"

nThread

The number of CPU cores you want to use to speed up your queries through parallelization.

conda_env

Conda environment to search for tabix executable in.

verbose

Show more (=TRUE) or fewer (=FALSE) messages.

See also