Merge files from a list of paths

Merge a list of files into one by stacking them on top of each other (i.e. rbind).

Usage

merge_files(file_paths, nThread = 1, verbose = TRUE)

Arguments

file_paths

Paths of files to import and merge into one data.table.

nThread

Number of threads to parallelize reading in files across.

verbose

Print messages.

See also

Examples

if (FALSE) { # \dontrun{
sumstats_paths <- echodata::get_Nalls2019_loci(limit_snps = 5)
merged_dat <- merge_files(file_paths = sumstats_paths)
} # }