Rapid querying, colocalization, and plotting of summary stats from the eQTL Catalogue

The functions in catalogueR are partly derived from the following eQTL Catalogue tutorial. Additional eQTL Catalogue Resources:

GitHub

In-depth API documentation

FTP server

Details

Notes on parallelization: There's multiple levels to parallelize on. You can only chooose one level at a time:

multithread_qtl=T

Across QTL datasets

multithread_loci=T

Across loci

multithread_tabix=T

Within tabix files

You can also get a speedup by using tabix instead of the rest API Test: For 3 loci, and X QTL datasets:

RESTful API:

7.5 minutes

Tabix:

27 seconds (*clear winner! ~17x speedup)

That said, if you're only query a small number of specific SNPs (rather than a large range), the RESTful API can sometimes be faster.

See also

Author

Maintainer: Brian Schilder brian_schilder@alumni.brown.edu

Other contributors:

Jack Humphrey [contributor]