Iterate queries to eQTL Catalogue

Deetermines which datasets to query using qtl_search. Uses coordinates from stored summary stats files (e.g. GWAS) to determine which regions to query from eQTL Catalogue. Each locus file can be stored separately, or merged together to form one large file with all query results.

eQTL_Catalogue.query(
  sumstats_paths = NULL,
  output_dir = "./catalogueR_queries",
  qtl_search = NULL,
  use_tabix = T,
  nThread = 4,
  quant_method = "ge",
  infer_region = T,
  split_files = T,
  merge_with_gwas = T,
  force_new_subset = F,
  genome_build = "hg19",
  progress_bar = T,
  verbose = T
)

Arguments

sumstats_paths
A list of paths to any number of summary stats files whose coordinates you want to use to make queries to eQTL Catalogue. If you wish to add custom names to the loci, simply add these as the names of the path list (e.g. c(BST1="<path>/<to>/<BST1_file>", LRRK2="<path>/<to>/<LRRK2_file>")). Otherwise, loci will automatically named based on their min/max genomic coordinates.

The minimum columns in these files required to make queries include:

SNP

RSID of each SNP.

CHR

Chromosome (can be in "chr12" or "12" format).

POS

Genomic position of each SNP.

...

Optional extra columns.

output_dir
The folder you want the merged gwas/qtl results to be saved to (set output_dir=F if you don't want to save the results). If split_files=F, all query results will be merged into one and saved as <output_dir>/eQTL_Catalogue.tsv.gz. If split_files=T, all query results will instead be split into smaller files and stored in <output_dir>/.

qtl_search
This function will automatically search for any datasets that match your criterion. For example, if you search "Alasoo_2018", it will query the datasets:

Alasoo_2018.macrophage_naive

Alasoo_2018.macrophage_Salmonella

Alasoo_2018.macrophage_IFNg+Salmonella

You can be more specific about which datasets you want to include, for example by searching: "Alasoo_2018.macrophage_IFNg". You can even search by tissue or condition type (e.g.c("blood","brain")) and any QTL datasets containing those substrings (case-insensitive) in their name or metadata will be queried too.

use_tabix
Tabix is about ~17x faster (default: =T) than the REST API (=F).

nThread
The number of CPU cores you want to use to speed up your queries through parallelization.

quant_method
eQTL Catalogue actually contains more than just eQTL data. For each dataset, the following kinds of QTLs can be queried:

gene expression QTL

quant_method="ge" (default) or quant_method="microarray", depending on the dataset. catalogueR will automatically select whichever option is available.

exon expression QTL

*under construction* quant_method="ex"

transcript usage QTL

*under construction* quant_method="tx"

promoter, splice junction and 3' end usage QTL

*under construction* quant_method="txrev"

split_files
Save the results as one file per QTL dataset (with all loci within each file). If this is set to =T, then this function will return the list of paths where these files were saved. A helper function is provided to import and merge them back together in R. If this is set to =F, then this function will instead return one big merged data.table containing results from all QTL datasets and all loci. =F is not recommended when you have many large loci and/or many QTL datasets, because you can only fit so much data into memory.

merge_with_gwas
Whether you want to merge your QTL query results with your GWAS data (convenient, but takes up more storage).

force_new_subset
By default, catalogueR will use any pre-existing files that match your query. Set force_new_subset=T to override this and force a new query.

genome_build
The genome build of your query coordinates (e.g. gwas_data). If your coordinates are in hg19, catalogueR will automatically lift them over to hg38 (as this is the build that eQTL Catalogue uses).

progress_bar
progress_bar=T allows progress to be monitored even when multithreading enabled. Requires R package pbmcapply.

verbose
Show more (=T) or fewer (=F) messages.

See also

Examples

sumstats_paths <- example_sumstats_paths()

# Merged results
GWAS.QTL <- eQTL_Catalogue.query(sumstats_paths=sumstats_paths, qtl_search="Alasoo_2018", nThread=1, force_new_subset=T, merge_with_gwas=F, progress_bar=T, split_files=F)

#> [1] "+ Optimizing multi-threading..."
#> [1] "++ Multi-threading across QTL datasets."
#> [1] "eQTL_Catalogue:: Querying 4 QTL datasets x 3 GWAS loci (12 total)"

#> Alasoo_2018.macrophage_naive

#> [1] "++ Extracting locus name from `sumstats_paths` names."

#> _+_+_+_+_+_+_+_+_--- Locus: BST1 ---_+_+_+_+_+_+_+_+_

#> Start at 2020-07-05 17:25:44

#> 

#> 
#> End at 2020-07-05 17:25:48

#> Runtime in total is: 4 secs

#> [1] "++ Inferring coordinates from gwas_data"
#> [1] "eQTL_Catalogue:: 78481 SNPs returned in 18.9 seconds."
#> [1] "++ Converting: Ensembl IDs ==> HGNC gene symbols"
#> [1] "++ Adding `Locus.GWAS` column."
#> [1] "++ Extracting locus name from `sumstats_paths` names."

#> _+_+_+_+_+_+_+_+_--- Locus: LRRK2 ---_+_+_+_+_+_+_+_+_

#> Start at 2020-07-05 17:26:07

#> 

#> 
#> End at 2020-07-05 17:26:13

#> Runtime in total is: 6 secs

#> [1] "++ Inferring coordinates from gwas_data"
#> [1] "eQTL_Catalogue:: 2795 SNPs returned in 5.8 seconds."
#> [1] "++ Converting: Ensembl IDs ==> HGNC gene symbols"
#> [1] "++ Adding `Locus.GWAS` column."
#> [1] "++ Extracting locus name from `sumstats_paths` names."

#> _+_+_+_+_+_+_+_+_--- Locus: MEX3C ---_+_+_+_+_+_+_+_+_

#> Start at 2020-07-05 17:26:19

#> 

#> 
#> End at 2020-07-05 17:26:24

#> Runtime in total is: 5 secs

#> [1] "++ Inferring coordinates from gwas_data"
#> [1] "eQTL_Catalogue:: 50088 SNPs returned in 9.4 seconds."
#> [1] "++ Converting: Ensembl IDs ==> HGNC gene symbols"
#> [1] "++ Adding `Locus.GWAS` column."
#> [1] "+ Returning merged data.table of query results."

#> Alasoo_2018.macrophage_IFNg

#> [1] "++ Extracting locus name from `sumstats_paths` names."

#> _+_+_+_+_+_+_+_+_--- Locus: BST1 ---_+_+_+_+_+_+_+_+_

#> Start at 2020-07-05 17:26:33

#> 

#> 
#> End at 2020-07-05 17:26:37

#> Runtime in total is: 4 secs

#> [1] "++ Inferring coordinates from gwas_data"
#> [1] "eQTL_Catalogue:: 57036 SNPs returned in 11.2 seconds."
#> [1] "++ Converting: Ensembl IDs ==> HGNC gene symbols"
#> [1] "++ Adding `Locus.GWAS` column."
#> [1] "++ Extracting locus name from `sumstats_paths` names."

#> _+_+_+_+_+_+_+_+_--- Locus: LRRK2 ---_+_+_+_+_+_+_+_+_

#> Start at 2020-07-05 17:26:48

#> 

#> 
#> End at 2020-07-05 17:26:53

#> Runtime in total is: 5 secs

#> [1] "++ Inferring coordinates from gwas_data"
#> [1] "eQTL_Catalogue:: 3354 SNPs returned in 6.2 seconds."
#> [1] "++ Converting: Ensembl IDs ==> HGNC gene symbols"
#> [1] "++ Adding `Locus.GWAS` column."
#> [1] "++ Extracting locus name from `sumstats_paths` names."

#> _+_+_+_+_+_+_+_+_--- Locus: MEX3C ---_+_+_+_+_+_+_+_+_

#> Start at 2020-07-05 17:26:59

#> 

#> 
#> End at 2020-07-05 17:27:02

#> Runtime in total is: 3 secs

#> [1] "++ Inferring coordinates from gwas_data"
#> [1] "eQTL_Catalogue:: 51536 SNPs returned in 9.7 seconds."
#> [1] "++ Converting: Ensembl IDs ==> HGNC gene symbols"
#> [1] "++ Adding `Locus.GWAS` column."
#> [1] "+ Returning merged data.table of query results."

#> Alasoo_2018.macrophage_Salmonella

#> [1] "++ Extracting locus name from `sumstats_paths` names."

#> _+_+_+_+_+_+_+_+_--- Locus: BST1 ---_+_+_+_+_+_+_+_+_

#> Start at 2020-07-05 17:27:12

#> 

#> 
#> End at 2020-07-05 17:27:17

#> Runtime in total is: 5 secs

#> [1] "++ Inferring coordinates from gwas_data"
#> [1] "eQTL_Catalogue:: 68136 SNPs returned in 21.7 seconds."
#> [1] "++ Converting: Ensembl IDs ==> HGNC gene symbols"
#> [1] "++ Adding `Locus.GWAS` column."
#> [1] "++ Extracting locus name from `sumstats_paths` names."

#> _+_+_+_+_+_+_+_+_--- Locus: LRRK2 ---_+_+_+_+_+_+_+_+_

#> Start at 2020-07-05 17:27:38

#> 

#> 
#> End at 2020-07-05 17:27:42

#> Runtime in total is: 4 secs

#> [1] "++ Inferring coordinates from gwas_data"
#> [1] "eQTL_Catalogue:: 2795 SNPs returned in 5.4 seconds."
#> [1] "++ Converting: Ensembl IDs ==> HGNC gene symbols"
#> [1] "++ Adding `Locus.GWAS` column."
#> [1] "++ Extracting locus name from `sumstats_paths` names."

#> _+_+_+_+_+_+_+_+_--- Locus: MEX3C ---_+_+_+_+_+_+_+_+_

#> Start at 2020-07-05 17:27:48

#> 

#> 
#> End at 2020-07-05 17:27:53

#> Runtime in total is: 5 secs

#> [1] "++ Inferring coordinates from gwas_data"
#> [1] "eQTL_Catalogue:: 51536 SNPs returned in 9.2 seconds."
#> [1] "++ Converting: Ensembl IDs ==> HGNC gene symbols"
#> [1] "++ Adding `Locus.GWAS` column."
#> [1] "+ Returning merged data.table of query results."

#> Alasoo_2018.macrophage_IFNg+Salmonella

#> [1] "++ Extracting locus name from `sumstats_paths` names."

#> _+_+_+_+_+_+_+_+_--- Locus: BST1 ---_+_+_+_+_+_+_+_+_

#> Start at 2020-07-05 17:28:02

#> 

#> 
#> End at 2020-07-05 17:28:08

#> Runtime in total is: 6 secs

#> [1] "++ Inferring coordinates from gwas_data"
#> [1] "eQTL_Catalogue:: 43983 SNPs returned in 14.8 seconds."
#> [1] "++ Converting: Ensembl IDs ==> HGNC gene symbols"
#> [1] "++ Adding `Locus.GWAS` column."
#> [1] "++ Extracting locus name from `sumstats_paths` names."

#> _+_+_+_+_+_+_+_+_--- Locus: LRRK2 ---_+_+_+_+_+_+_+_+_

#> Start at 2020-07-05 17:28:23

#> 

#> 
#> End at 2020-07-05 17:28:28

#> Runtime in total is: 5 secs

#> [1] "++ Inferring coordinates from gwas_data"
#> [1] "eQTL_Catalogue:: 3354 SNPs returned in 5.6 seconds."
#> [1] "++ Converting: Ensembl IDs ==> HGNC gene symbols"
#> [1] "++ Adding `Locus.GWAS` column."
#> [1] "++ Extracting locus name from `sumstats_paths` names."

#> _+_+_+_+_+_+_+_+_--- Locus: MEX3C ---_+_+_+_+_+_+_+_+_

#> Start at 2020-07-05 17:28:33

#> 

#> 
#> End at 2020-07-05 17:28:37

#> Runtime in total is: 4 secs

#> [1] "++ Inferring coordinates from gwas_data"
#> [1] "eQTL_Catalogue:: 45995 SNPs returned in 7.4 seconds."
#> [1] "++ Converting: Ensembl IDs ==> HGNC gene symbols"
#> [1] "++ Adding `Locus.GWAS` column."
#> [1] "+ Returning merged data.table of query results."
#> [1] "++ Post-processing merged results."
#> [1] "+ Checking results for missing queries..."
#> [1] "+ Saving merged query results ==> ./catalogueR_queries/eQTL_Catalogue.tsv.gz"
#> Time difference of 3.1 mins

# Merged results (parallel)
GWAS.QTL <- eQTL_Catalogue.query(sumstats_paths=sumstats_paths, qtl_search="Alasoo_2018", nThread=4, force_new_subset=T, merge_with_gwas=F, progress_bar=T, split_files=F)

#> [1] "+ Optimizing multi-threading..."
#> [1] "++ Multi-threading across QTL datasets."
#> [1] "eQTL_Catalogue:: Querying 4 QTL datasets x 3 GWAS loci (12 total)"
#> [1] "++ Post-processing merged results."
#> [1] "+ Checking results for missing queries..."
#> [1] "+ QTL datasets with no hits/failed to pull data:"
#> [1] "  + BST1__Alasoo_2018.macrophage_naive"
#> [1] "  + LRRK2__Alasoo_2018.macrophage_naive"
#> [1] "  + MEX3C__Alasoo_2018.macrophage_naive"
#> [1] "  + BST1__Alasoo_2018.macrophage_IFNg"
#> [1] "  + LRRK2__Alasoo_2018.macrophage_IFNg"
#> [1] "  + MEX3C__Alasoo_2018.macrophage_IFNg"
#> [1] "  + BST1__Alasoo_2018.macrophage_Salmonella"
#> [1] "  + LRRK2__Alasoo_2018.macrophage_Salmonella"
#> [1] "  + MEX3C__Alasoo_2018.macrophage_Salmonella"
#> [1] "+ Saving merged query results ==> ./catalogueR_queries/eQTL_Catalogue.tsv.gz"
#> Time difference of 57.5 secs

# Split results
gwas.qtl <- eQTL_Catalogue.query(sumstats_paths=sumstats_paths, qtl_search="Alasoo_2018", nThread=1, force_new_subset=T, merge_with_gwas=F, progress_bar=T)

#> [1] "+ Optimizing multi-threading..."
#> [1] "++ Multi-threading across QTL datasets."
#> [1] "eQTL_Catalogue:: Querying 4 QTL datasets x 3 GWAS loci (12 total)"

#> Alasoo_2018.macrophage_naive

#> [1] "++ Extracting locus name from `sumstats_paths` names."

#> _+_+_+_+_+_+_+_+_--- Locus: BST1 ---_+_+_+_+_+_+_+_+_

#> Start at 2020-07-05 17:29:45

#> 

#> 
#> End at 2020-07-05 17:29:48

#> Runtime in total is: 3 secs

#> [1] "++ Inferring coordinates from gwas_data"
#> [1] "eQTL_Catalogue:: 78481 SNPs returned in 18.3 seconds."
#> [1] "++ Converting: Ensembl IDs ==> HGNC gene symbols"
#> [1] "++ Adding `Locus.GWAS` column."
#> [1] "++ Saving split file ==> ./catalogueR_queries/Alasoo_2018.macrophage_naive/BST1__Alasoo_2018.macrophage_naive.tsv.gz"
#> [1] "++ Extracting locus name from `sumstats_paths` names."

#> _+_+_+_+_+_+_+_+_--- Locus: LRRK2 ---_+_+_+_+_+_+_+_+_

#> Start at 2020-07-05 17:30:06

#> 

#> 
#> End at 2020-07-05 17:30:11

#> Runtime in total is: 5 secs

#> [1] "++ Inferring coordinates from gwas_data"
#> [1] "eQTL_Catalogue:: 2795 SNPs returned in 5.8 seconds."
#> [1] "++ Converting: Ensembl IDs ==> HGNC gene symbols"
#> [1] "++ Adding `Locus.GWAS` column."
#> [1] "++ Saving split file ==> ./catalogueR_queries/Alasoo_2018.macrophage_naive/LRRK2__Alasoo_2018.macrophage_naive.tsv.gz"
#> [1] "++ Extracting locus name from `sumstats_paths` names."

#> _+_+_+_+_+_+_+_+_--- Locus: MEX3C ---_+_+_+_+_+_+_+_+_

#> Start at 2020-07-05 17:30:17

#> 

#> 
#> End at 2020-07-05 17:30:20

#> Runtime in total is: 3 secs

#> [1] "++ Inferring coordinates from gwas_data"
#> [1] "eQTL_Catalogue:: 50088 SNPs returned in 8.8 seconds."
#> [1] "++ Converting: Ensembl IDs ==> HGNC gene symbols"
#> [1] "++ Adding `Locus.GWAS` column."
#> [1] "++ Saving split file ==> ./catalogueR_queries/Alasoo_2018.macrophage_naive/MEX3C__Alasoo_2018.macrophage_naive.tsv.gz"
#> [1] "+ Returning list of split query results files."

#> Alasoo_2018.macrophage_IFNg

#> [1] "++ Extracting locus name from `sumstats_paths` names."

#> _+_+_+_+_+_+_+_+_--- Locus: BST1 ---_+_+_+_+_+_+_+_+_

#> Start at 2020-07-05 17:30:29

#> 

#> 
#> End at 2020-07-05 17:30:34

#> Runtime in total is: 5 secs

#> [1] "++ Inferring coordinates from gwas_data"
#> [1] "eQTL_Catalogue:: 57036 SNPs returned in 16.1 seconds."
#> [1] "++ Converting: Ensembl IDs ==> HGNC gene symbols"
#> [1] "++ Adding `Locus.GWAS` column."
#> [1] "++ Saving split file ==> ./catalogueR_queries/Alasoo_2018.macrophage_IFNg/BST1__Alasoo_2018.macrophage_IFNg.tsv.gz"
#> [1] "++ Extracting locus name from `sumstats_paths` names."

#> _+_+_+_+_+_+_+_+_--- Locus: LRRK2 ---_+_+_+_+_+_+_+_+_

#> Start at 2020-07-05 17:30:50

#> 

#> 
#> End at 2020-07-05 17:30:54

#> Runtime in total is: 4 secs

#> [1] "++ Inferring coordinates from gwas_data"
#> [1] "eQTL_Catalogue:: 3354 SNPs returned in 5.8 seconds."
#> [1] "++ Converting: Ensembl IDs ==> HGNC gene symbols"
#> [1] "++ Adding `Locus.GWAS` column."
#> [1] "++ Saving split file ==> ./catalogueR_queries/Alasoo_2018.macrophage_IFNg/LRRK2__Alasoo_2018.macrophage_IFNg.tsv.gz"
#> [1] "++ Extracting locus name from `sumstats_paths` names."

#> _+_+_+_+_+_+_+_+_--- Locus: MEX3C ---_+_+_+_+_+_+_+_+_

#> Start at 2020-07-05 17:31:00

#> 

#> 
#> End at 2020-07-05 17:31:03

#> Runtime in total is: 3 secs

#> [1] "++ Inferring coordinates from gwas_data"
#> [1] "eQTL_Catalogue:: 51536 SNPs returned in 12.1 seconds."
#> [1] "++ Converting: Ensembl IDs ==> HGNC gene symbols"
#> [1] "++ Adding `Locus.GWAS` column."
#> [1] "++ Saving split file ==> ./catalogueR_queries/Alasoo_2018.macrophage_IFNg/MEX3C__Alasoo_2018.macrophage_IFNg.tsv.gz"
#> [1] "+ Returning list of split query results files."

#> Alasoo_2018.macrophage_Salmonella

#> [1] "++ Extracting locus name from `sumstats_paths` names."

#> _+_+_+_+_+_+_+_+_--- Locus: BST1 ---_+_+_+_+_+_+_+_+_

#> Start at 2020-07-05 17:31:16

#> 

#> 
#> End at 2020-07-05 17:31:18

#> Runtime in total is: 2 secs

#> [1] "++ Inferring coordinates from gwas_data"
#> [1] "eQTL_Catalogue:: 68136 SNPs returned in 14.9 seconds."
#> [1] "++ Converting: Ensembl IDs ==> HGNC gene symbols"
#> [1] "++ Adding `Locus.GWAS` column."
#> [1] "++ Saving split file ==> ./catalogueR_queries/Alasoo_2018.macrophage_Salmonella/BST1__Alasoo_2018.macrophage_Salmonella.tsv.gz"
#> [1] "++ Extracting locus name from `sumstats_paths` names."

#> _+_+_+_+_+_+_+_+_--- Locus: LRRK2 ---_+_+_+_+_+_+_+_+_

#> Start at 2020-07-05 17:31:34

#> 

#> 
#> End at 2020-07-05 17:31:38

#> Runtime in total is: 4 secs

#> [1] "++ Inferring coordinates from gwas_data"
#> [1] "eQTL_Catalogue:: 2795 SNPs returned in 6.1 seconds."
#> [1] "++ Converting: Ensembl IDs ==> HGNC gene symbols"
#> [1] "++ Adding `Locus.GWAS` column."
#> [1] "++ Saving split file ==> ./catalogueR_queries/Alasoo_2018.macrophage_Salmonella/LRRK2__Alasoo_2018.macrophage_Salmonella.tsv.gz"
#> [1] "++ Extracting locus name from `sumstats_paths` names."

#> _+_+_+_+_+_+_+_+_--- Locus: MEX3C ---_+_+_+_+_+_+_+_+_

#> Start at 2020-07-05 17:31:44

#> 

#> 
#> End at 2020-07-05 17:31:48

#> Runtime in total is: 4 secs

#> [1] "++ Inferring coordinates from gwas_data"
#> [1] "eQTL_Catalogue:: 51536 SNPs returned in 6.9 seconds."
#> [1] "++ Converting: Ensembl IDs ==> HGNC gene symbols"
#> [1] "++ Adding `Locus.GWAS` column."
#> [1] "++ Saving split file ==> ./catalogueR_queries/Alasoo_2018.macrophage_Salmonella/MEX3C__Alasoo_2018.macrophage_Salmonella.tsv.gz"
#> [1] "+ Returning list of split query results files."

#> Alasoo_2018.macrophage_IFNg+Salmonella

#> [1] "++ Extracting locus name from `sumstats_paths` names."

#> _+_+_+_+_+_+_+_+_--- Locus: BST1 ---_+_+_+_+_+_+_+_+_

#> Start at 2020-07-05 17:31:55

#> 

#> 
#> End at 2020-07-05 17:32:00

#> Runtime in total is: 5 secs

#> [1] "++ Inferring coordinates from gwas_data"
#> [1] "eQTL_Catalogue:: 43983 SNPs returned in 16.3 seconds."
#> [1] "++ Converting: Ensembl IDs ==> HGNC gene symbols"
#> [1] "++ Adding `Locus.GWAS` column."
#> [1] "++ Saving split file ==> ./catalogueR_queries/Alasoo_2018.macrophage_IFNg+Salmonella/BST1__Alasoo_2018.macrophage_IFNg+Salmonella.tsv.gz"
#> [1] "++ Extracting locus name from `sumstats_paths` names."

#> _+_+_+_+_+_+_+_+_--- Locus: LRRK2 ---_+_+_+_+_+_+_+_+_

#> Start at 2020-07-05 17:32:16

#> 

#> 
#> End at 2020-07-05 17:32:20

#> Runtime in total is: 4 secs

#> [1] "++ Inferring coordinates from gwas_data"
#> [1] "eQTL_Catalogue:: 3354 SNPs returned in 5.8 seconds."
#> [1] "++ Converting: Ensembl IDs ==> HGNC gene symbols"
#> [1] "++ Adding `Locus.GWAS` column."
#> [1] "++ Saving split file ==> ./catalogueR_queries/Alasoo_2018.macrophage_IFNg+Salmonella/LRRK2__Alasoo_2018.macrophage_IFNg+Salmonella.tsv.gz"
#> [1] "++ Extracting locus name from `sumstats_paths` names."

#> _+_+_+_+_+_+_+_+_--- Locus: MEX3C ---_+_+_+_+_+_+_+_+_

#> Start at 2020-07-05 17:32:26

#> 

#> 
#> End at 2020-07-05 17:32:29

#> Runtime in total is: 3 secs

#> [1] "++ Inferring coordinates from gwas_data"
#> [1] "eQTL_Catalogue:: 45995 SNPs returned in 7.4 seconds."
#> [1] "++ Converting: Ensembl IDs ==> HGNC gene symbols"
#> [1] "++ Adding `Locus.GWAS` column."
#> [1] "++ Saving split file ==> ./catalogueR_queries/Alasoo_2018.macrophage_IFNg+Salmonella/MEX3C__Alasoo_2018.macrophage_IFNg+Salmonella.tsv.gz"
#> [1] "+ Returning list of split query results files."
#> [1] "++ Returning list of split files paths."
#> Time difference of 2.9 mins

GWAS.QTL <- gather_files(file_paths = gwas.qtl)

#> [1] "+ Merging 12 files."
#> [1] "+ Using 4 cores."
#> [1] "+ Merged data.table: 459089 rows x 21 columns."

# Split results (parallel)
gwas.qtl <- eQTL_Catalogue.query(sumstats_paths=sumstats_paths, qtl_search="Alasoo_2018", nThread=4, force_new_subset=T, merge_with_gwas=F, progress_bar=T)

#> [1] "+ Optimizing multi-threading..."
#> [1] "++ Multi-threading across QTL datasets."
#> [1] "eQTL_Catalogue:: Querying 4 QTL datasets x 3 GWAS loci (12 total)"
#> [1] "++ Returning list of split files paths."
#> Time difference of 51.3 secs

GWAS.QTL <- gather_files(file_paths = gwas.qtl)

#> [1] "+ Merging 3 files."
#> [1] "+ Using 4 cores."
#> [1] "+ Merged data.table: 131364 rows x 21 columns."